+++ Assets/OxGFrame/AssetLoader/Scripts/Runtime/Bundle/BundleConfig.cs
@@ -27,7 +27,8 @@ namespace OxGFrame.AssetLoader.Bundle
         {
             EditorSimulateMode,
             OfflineMode,
-            HostMode
+            HostMode,
+            WebGLMode
         }
 
         public class CryptogramType