+++ Assets/OxGFrame/AssetLoader/Scripts/Runtime/Bundle/YooAssets/QueryServices.cs
@@ -1,40 +1,40 @@
-﻿using System.IO;
+﻿using System.Collections;
+using System.IO;
 using UnityEngine;
 using UnityEngine.Networking;
 using YooAsset;
 
 namespace OxGFrame.AssetLoader.Bundle
 {
-    public class RequestBuiltinQuery : IQueryServices
+    public abstract class RequestQueryBase : IQueryServices
     {
-        public bool QueryStreamingAssets(string packageName, string fileName, bool isRawFile)
+        public DeliveryFileInfo GetDeliveryFileInfo(string packageName, string fileName)
         {
-            bool result;
+            throw new System.NotImplementedException();
+        }
 
-            string builtinPackagePath = BundleConfig.GetBuiltinPackagePath(packageName);
-            string url = Path.Combine(builtinPackagePath, fileName);
+        public bool QueryDeliveryFiles(string packageName, string fileName)
+        {
+            return false;
+        }
 
-            UnityWebRequest request = null;
-            try
-            {
-                request = UnityWebRequest.Get(url);
-                request.SendWebRequest();
-            }
-            catch
-            {
-                Debug.Log($"<color=#FF0000>Request failed. URL: {url}</color>");
-                request?.Dispose();
-                result = false;
-                return result;
-            }
+        public virtual bool QueryStreamingAssets(string packageName, string fileName, bool isRawFile)
+        {
+            return false;
+        }
+
+        protected IEnumerator WebRequest(string url)
+        {
+            UnityWebRequest request = UnityWebRequest.Get(url);
+            request.downloadHandler = new DownloadHandlerBuffer();
+            request.SendWebRequest();
 
             while (true)
             {
                 if (request == null)
                 {
                     Debug.Log($"<color=#FF0000>Request is null. URL: {url}</color>");
-                    result = false;
-                    break;
+                    yield return false;
                 }
 
                 // 網路中斷後主線程異常, 編輯器下無法正常運行, 不過發佈模式沒問題
@@ -44,29 +44,51 @@ namespace OxGFrame.AssetLoader.Bundle
                 {
                     Debug.Log($"<color=#FF0000>【Query】Request failed. Code: {request.responseCode}, URL: {url}</color>");
                     request?.Dispose();
-                    result = false;
-                    break;
+                    yield return false;
                 }
 
                 if (request.downloadedBytes > 0 && request.responseCode < 400)
                 {
                     Debug.Log($"<color=#00FF00>【Query】Request succeeded. Code: {request.responseCode}, PartialBytes: {request.downloadedBytes}, URL: {url}</color>");
                     request?.Dispose();
-                    result = true;
-                    break;
+                    yield return true;
                 }
-            }
 
-            return result;
+                yield return null;
+            }
         }
     }
 
-    public class RequestSandboxQuery : IQueryServices
+    public class RequestBuiltinQuery : RequestQueryBase
     {
-        public bool QueryStreamingAssets(string packageName, string fileName, bool isRawFile)
+        public override bool QueryStreamingAssets(string packageName, string fileName, bool isRawFile)
         {
-            bool result;
+#if UNITY_WEBGL
+            return StreamingAssetsHelper.FileExists(packageName, fileName);
+#else
+            #region Builtin (StreamingAssets)
+            string builtinPackagePath = BundleConfig.GetBuiltinPackagePath(packageName);
+            string url = Path.Combine(builtinPackagePath, fileName);
+            #endregion
 
+            var e = this.WebRequest(url);
+            while (e.MoveNext())
+                if (e.Current != null)
+                    return (bool)e.Current;
+
+            return false;
+#endif
+        }
+    }
+
+    public class RequestSandboxQuery : RequestQueryBase
+    {
+        public override bool QueryStreamingAssets(string packageName, string fileName, bool isRawFile)
+        {
+#if UNITY_WEBGL
+            return false;
+#else
+            #region Sandbox (PersistentData)
             string sandboxPackagePath = BundleConfig.GetLocalSandboxPackagePath(packageName);
             string hashPrefix = fileName.Substring(0, 2);
             string hashFolderName = fileName.Split(".")[0];
@@ -80,51 +102,15 @@ namespace OxGFrame.AssetLoader.Bundle
                 hashFolderName,
                 BundleConfig.yooBundleFileName
             );
+            #endregion
 
-            UnityWebRequest request = null;
-            try
-            {
-                request = UnityWebRequest.Get(url);
-                request.SendWebRequest();
-            }
-            catch
-            {
-                Debug.Log($"<color=#FF0000>Request failed. URL: {url}</color>");
-                request?.Dispose();
-                result = false;
-                return result;
-            }
+            var e = this.WebRequest(url);
+            while (e.MoveNext())
+                if (e.Current != null)
+                    return (bool)e.Current;
 
-            while (true)
-            {
-                if (request == null)
-                {
-                    Debug.Log($"<color=#FF0000>Request is null. URL: {url}</color>");
-                    result = false;
-                    break;
-                }
-
-                // 網路中斷後主線程異常, 編輯器下無法正常運行, 不過發佈模式沒問題
-                if (request.result == UnityWebRequest.Result.DataProcessingError ||
-                    request.result == UnityWebRequest.Result.ProtocolError ||
-                    request.result == UnityWebRequest.Result.ConnectionError)
-                {
-                    Debug.Log($"<color=#FF0000>【Query】Request failed. Code: {request.responseCode}, URL: {url}</color>");
-                    request?.Dispose();
-                    result = false;
-                    break;
-                }
-
-                if (request.downloadedBytes > 0 && request.responseCode < 400)
-                {
-                    Debug.Log($"<color=#00FF00>【Query】Request succeeded. Code: {request.responseCode}, PartialBytes: {request.downloadedBytes}, URL: {url}</color>");
-                    request?.Dispose();
-                    result = true;
-                    break;
-                }
-            }
-
-            return result;
+            return false;
+#endif
         }
     }
 }