+++ Assets/OxGFrame/AssetLoader/Scripts/Runtime/Bundle/YooAssets/StreamingAssetsHelper/BuildinFileManifest.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+using UnityEngine;
+
+namespace OxGFrame.AssetLoader.Bundle
+{
+    // Reference: YooAsset
+
+    /// <summary>
+    /// 内置资源清单
+    /// </summary>
+    public class BuildinFileManifest : ScriptableObject
+    {
+        public List<string> BuildinFiles = new List<string>();
+    }
+}