+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Runtime/PackageSystem/PlayMode/HostPlayModeImpl.cs
@@ -51,6 +51,24 @@ namespace YooAsset
 			return bundleInfo;
 		}
 
+		// 查询相关
+		private bool IsBuildinPackageBundle(PackageBundle packageBundle)
+		{
+			return _queryServices.QueryStreamingAssets(_packageName, packageBundle.FileName, packageBundle.IsRawFile);
+		}
+		private bool IsCachedPackageBundle(PackageBundle packageBundle)
+		{
+			return CacheSystem.IsCached(packageBundle.PackageName, packageBundle.CacheGUID);
+		}
+		private bool IsDeliveryPackageBundle(PackageBundle packageBundle)
+		{
+			return _queryServices.QueryDeliveryFiles(_packageName, packageBundle.FileName);
+		}
+		private DeliveryFileInfo GetDeiveryFileInfo(PackageBundle packageBundle)
+		{
+			return _queryServices.GetDeliveryFileInfo(_packageName, packageBundle.FileName);
+		}
+
 		#region IPlayModeServices接口
 		public PackageManifest ActiveManifest
 		{
@@ -69,15 +87,6 @@ namespace YooAsset
 			{
 				PersistentTools.GetPersistent(_packageName).SaveSandboxPackageVersionFile(_activeManifest.PackageVersion);
 			}
-		}
-
-		private bool IsBuildinPackageBundle(PackageBundle packageBundle)
-		{
-			return _queryServices.QueryStreamingAssets(_packageName, packageBundle.FileName, packageBundle.IsRawFile);
-		}
-		private bool IsCachedPackageBundle(PackageBundle packageBundle)
-		{
-			return CacheSystem.IsCached(packageBundle.PackageName, packageBundle.CacheGUID);
 		}
 
 		UpdatePackageVersionOperation IPlayModeServices.UpdatePackageVersionAsync(bool appendTimeTicks, int timeout)
@@ -110,6 +119,10 @@ namespace YooAsset
 			List<PackageBundle> downloadList = new List<PackageBundle>(1000);
 			foreach (var packageBundle in manifest.BundleList)
 			{
+				// 忽略分发文件
+				if (IsDeliveryPackageBundle(packageBundle))
+					continue;
+
 				// 忽略缓存文件
 				if (IsCachedPackageBundle(packageBundle))
 					continue;
@@ -135,6 +148,10 @@ namespace YooAsset
 			List<PackageBundle> downloadList = new List<PackageBundle>(1000);
 			foreach (var packageBundle in manifest.BundleList)
 			{
+				// 忽略分发文件
+				if (IsDeliveryPackageBundle(packageBundle))
+					continue;
+
 				// 忽略缓存文件
 				if (IsCachedPackageBundle(packageBundle))
 					continue;
@@ -196,6 +213,10 @@ namespace YooAsset
 			List<PackageBundle> downloadList = new List<PackageBundle>(1000);
 			foreach (var packageBundle in checkList)
 			{
+				// 忽略分发文件
+				if (IsDeliveryPackageBundle(packageBundle))
+					continue;
+
 				// 忽略缓存文件
 				if (IsCachedPackageBundle(packageBundle))
 					continue;
@@ -268,6 +289,14 @@ namespace YooAsset
 		{
 			if (packageBundle == null)
 				throw new Exception("Should never get here !");
+
+			// 查询分发资源
+			if (IsDeliveryPackageBundle(packageBundle))
+			{
+				DeliveryFileInfo deliveryFileInfo = GetDeiveryFileInfo(packageBundle);
+				BundleInfo bundleInfo = new BundleInfo(packageBundle, BundleInfo.ELoadMode.LoadFromDelivery, deliveryFileInfo.DeliveryFilePath, deliveryFileInfo.DeliveryFileOffset);
+				return bundleInfo;
+			}
 
 			// 查询沙盒资源
 			if (IsCachedPackageBundle(packageBundle))