+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Runtime/Services/IQueryServices.cs
@@ -1,11 +1,30 @@
 ﻿
 namespace YooAsset
 {
-	public interface IQueryServices
-	{
-		/// <summary>
-		/// 查询内置资源
-		/// </summary>
-		bool QueryStreamingAssets(string packageName, string fileName, bool isRawFile);
-	}
+    /// <summary>
+    /// 分发的资源信息
+    /// </summary>
+    public struct DeliveryFileInfo
+    {
+        public string DeliveryFilePath;
+        public ulong DeliveryFileOffset;
+    }
+
+    public interface IQueryServices
+    {
+        /// <summary>
+        /// 查询应用程序里的内置资源是否存在
+        /// </summary>
+        bool QueryStreamingAssets(string packageName, string fileName, bool isRawFile);
+
+        /// <summary>
+        /// 查询是否为开发者分发的资源
+        /// </summary>
+        bool QueryDeliveryFiles(string packageName, string fileName);
+
+        /// <summary>
+        /// 获取开发者分发的资源信息
+        /// </summary>
+        DeliveryFileInfo GetDeliveryFileInfo(string packageName, string fileName);
+    }
 }